refactor: createStakeholderApi exports server, renamed to stakeholderApi.ts
