+++ mock/stakeholderApi.ts
@@ -3,15 +3,17 @@ import { rest } from 'msw'
 import { setupServer } from 'msw/node'
 
 const server = setupServer(
-  rest.get('*/stakeholder', (req, res, ctx) =>
-    res(
+  rest.get('*/stakeholder', (req, res, ctx) => {
+    const id = req.url.searchParams.get('stakeholder_id')
+    if (!id) return res(ctx.status(500))
+    return res(
       ctx.status(200, 'Mock success'),
       ctx.json({
         ...data,
-        id: req.url.searchParams.get('stakeholder_id'),
+        id,
       }),
-    ),
-  ),
+    )
+  }),
 )
 
 export default server
