+++ server/utils/stakeholderApi.spec.ts
@@ -1,5 +1,5 @@
-import server from '@mock/createStakeholderApi'
 import data from '@mock/stakeholder.json'
+import server from '@mock/stakeholderApi'
 import { getStakeholder, getStakeholderMap } from './stakeholderApi'
 
 beforeAll(() => {
@@ -20,4 +20,10 @@ it('should get map name from stakeholder api', async () => {
   const stakeholderId = data.id
   const map = await getStakeholderMap(stakeholderId)
   expect(map).toBe(data.map)
+})
+
+it('should handle error', async () => {
+  const stakeholderId = ''
+  const map = await getStakeholderMap(stakeholderId)
+  expect(map).toBeUndefined()
 })
