+++ server/utils/stakeholderApi.ts
@@ -8,12 +8,17 @@ export async function getStakeholder(
 ): Promise<Stakeholder | void> {
   if (!stakeholderApiRoute)
     return console.error('env var: STAKEHOLDER_API_ROUTE is not set')
-  const res = await axios.get(stakeholderApiRoute, {
-    params: {
-      stakeholder_id: stakeholderId,
-    },
-  })
-  return res.data
+
+  try {
+    const res = await axios.get(stakeholderApiRoute, {
+      params: {
+        stakeholder_id: stakeholderId,
+      },
+    })
+    return res.data
+  } catch (err) {
+    return console.error((err as Error).message)
+  }
 }
 
 export async function getStakeholderMap(
